Added:Lombok library And Use Anotation @Data in entity class

+++ src/main/java/com/bridgelabz/SpringEmployeePayrollApp/repository/EmployeeRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepo extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepo extends JpaRepository<Employee, Long> {
 }
